# chili_app/openapi.json
{
  "components": {
    "schemas": {
      "AlertDetailResponse": {
        "description": "Expanded alert record used by alert and investigation views.",
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/AlertListItem"
          },
          "policy_citations": {
            "items": {
              "$ref": "#/components/schemas/PolicyCitation"
            },
            "title": "Policy Citations",
            "type": "array"
          },
          "related_entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Entity Ids",
            "type": "array"
          }
        },
        "required": [
          "alert"
        ],
        "title": "AlertDetailResponse",
        "type": "object"
      },
      "AlertListItem": {
        "description": "Summary alert row consumed by the analyst feed.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_label": {
            "title": "Entity Label",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "evidence_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Pack Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "acknowledged",
              "investigating",
              "resolved",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "knowledge_base_id",
          "entity_id",
          "entity_type",
          "entity_label",
          "severity",
          "status",
          "title",
          "reasoning",
          "confidence",
          "created_at"
        ],
        "title": "AlertListItem",
        "type": "object"
      },
      "AlertListResponse": {
        "description": "Paginated alert feed response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "page"
        ],
        "title": "AlertListResponse",
        "type": "object"
      },
...
      "ApiEnvelope": {
        "description": "Common status envelope for simple mutation responses.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "ok"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "ApiEnvelope",
        "type": "object"
      },
...
      "ClusterResult": {
        "description": "A single cluster summary returned from GNN analysis.",
        "properties": {
          "anomaly_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Anomaly Score",
            "type": "number"
          },
          "cluster_id": {
            "title": "Cluster Id",
            "type": "string"
          },
          "entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "required": [
          "cluster_id",
          "anomaly_score"
        ],
        "title": "ClusterResult",
        "type": "object"
      },
...
      "EntityTimeseriesResponse": {
        "description": "Timeseries payload for entity trend charts.",
        "properties": {
          "availability_status": {
            "default": "available",
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availability Status",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/EntityTimeseriesPointResponse"
            },
            "title": "Points",
            "type": "array"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          }
        },
        "required": [
          "entity_id",
          "metric_name"
        ],
        "title": "EntityTimeseriesResponse",
        "type": "object"
      },
...
      "GnnClusterResponse": {
        "description": "A list of clusters discovered for one knowledge base.",
        "properties": {
          "clusters": {
            "items": {
              "$ref": "#/components/schemas/ClusterResult"
            },
            "title": "Clusters",
            "type": "array"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          }
        },
        "required": [
          "knowledge_base_id"
        ],
        "title": "GnnClusterResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricTimeseriesResponse": {
        "description": "A range-bounded list of timeseries data points.",
        "properties": {
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TimeseriesPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "knowledge_base_id",
          "metric_name",
          "start",
          "end"
        ],
        "title": "MetricTimeseriesResponse",
        "type": "object"
      },
...
      "PageInfo": {
        "description": "Pagination metadata for collection responses.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "PolicyCitation": {
        "description": "Policy reference surfaced with evidence or alerts.",
        "properties": {
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "source_document_id": {
            "title": "Source Document Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "citation_id",
          "title",
          "excerpt",
          "source_document_id"
        ],
        "title": "PolicyCitation",
        "type": "object"
      },
...
      "RiskFactorResponse": {
        "description": "Frontend risk-factor breakdown.",
        "properties": {
          "contribution": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Contribution",
            "type": "number"
          },
          "factor_name": {
            "title": "Factor Name",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "required": [
          "factor_name",
          "contribution"
        ],
        "title": "RiskFactorResponse",
        "type": "object"
      },
      "RiskScore": {
        "description": "A single ranked risk score entry.",
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "overall_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Score",
            "type": "number"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "entity_type",
          "overall_score",
          "risk_level"
        ],
        "title": "RiskScore",
        "type": "object"
      },
      "RiskScoreListResponse": {
        "description": "A paginated, ranked list of risk scores.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RiskScore"
            },
            "title": "Items",
            "type": "array"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "knowledge_base_id",
          "total"
        ],
        "title": "RiskScoreListResponse",
        "type": "object"
      },
      "RiskScoreResponse": {
        "description": "Risk summary for one entity.",
        "properties": {
          "availability_status": {
            "default": "available",
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availability Status",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/RiskFactorResponse"
            },
            "title": "Factors",
            "type": "array"
          },
          "overall_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Score",
            "type": "number"
          },
          "risk_level": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          }
        },
        "required": [
          "entity_id",
          "overall_score",
          "risk_level"
        ],
        "title": "RiskScoreResponse",
        "type": "object"
      },
      "TimeseriesPoint": {
        "description": "A single data point in a returned time-series.",
        "properties": {
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "observed_at",
          "value"
        ],
        "title": "TimeseriesPoint",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend API gateway for the chiliAI Graph RAG analytics platform.",
    "title": "chiliAI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/alerts": {
      "get": {
        "description": "Return the alert feed in the api.contracts shape (items + page).",
        "operationId": "list_alerts_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "kb",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kb"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/alerts/{alert_id}": {
      "get": {
        "description": "Return one alert detail with related entities and policy citations.",
        "operationId": "get_alert_alerts__alert_id__get",
        "parameters": [
          {
            "description": "Alert identifier.",
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "description": "Alert identifier.",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/alerts/{alert_id}/acknowledge": {
      "post": {
        "description": "Acknowledge an alert; returns an ApiEnvelope status receipt.",
        "operationId": "acknowledge_alert_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "description": "Alert identifier.",
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "description": "Alert identifier.",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/analytics/gnn/clusters": {
      "get": {
        "description": "Return GNN-derived clusters for a knowledge base.\n\nReturns an empty list when the GNN capability is disabled in config.",
        "operationId": "list_gnn_clusters_analytics_gnn_clusters_get",
        "parameters": [
          {
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GnnClusterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Gnn Clusters",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/risk-scores": {
      "get": {
        "description": "Return ranked risk scores for entities in a knowledge base.",
        "operationId": "list_risk_scores_analytics_risk_scores_get",
        "parameters": [
          {
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "exclusiveMinimum": 0,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScoreListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Risk Scores",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/risk-scores/{entity_id}": {
      "get": {
        "description": "Return the risk score breakdown for one entity.",
        "operationId": "get_risk_score_analytics_risk_scores__entity_id__get",
        "parameters": [
          {
            "description": "Entity identifier.",
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "description": "Entity identifier.",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "Knowledge base identifier.",
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "description": "Knowledge base identifier.",
              "minLength": 1,
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Risk Score",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/timeseries": {
      "get": {
        "description": "Return data points for one metric over a bounded time range.",
        "operationId": "query_timeseries_analytics_timeseries_get",
        "parameters": [
          {
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metric",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricTimeseriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Timeseries",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/timeseries/{entity_id}": {
      "get": {
        "description": "Return chartable time-series points for one entity.",
        "operationId": "get_entity_timeseries_analytics_timeseries__entity_id__get",
        "parameters": [
          {
            "description": "Entity identifier.",
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "description": "Entity identifier.",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "Knowledge base identifier.",
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "description": "Knowledge base identifier.",
              "minLength": 1,
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityTimeseriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entity Timeseries",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}